feat/ui : Side Note List DIR 구조 수정
