+++ src/frontend/src/App.css
@@ -182,7 +182,7 @@ table, th, td {
 }
 
 .noteList-nav h3 {
-    font-size: 12px;
+    font-size: 11px;
     font-family: "Roboto", sans-serif;
     font-weight: 600;
     color: grey;
